Update the namespace of PricingMatrix
Currently the namespace of the `PricingMatrix` is `platformCore`,
however the webservices API expects this namespace to be `listAcct`. It
also expects all of the records within the `PricingMatrix` to have the
namespace of `listAcct`. In order to remedy this we will now update the
namespace of the `PricingMatrix` to `listAcct` and also create records
for `Price`, PriceList`, and `Pricing`. This way they can all have the
correct namesapce of `listAcct`.